bldchn, checking for socklen_t more safe
This patch is needed because currently we fail always where both
ws2tcpip.h and socklen_t is needed.

Note this patch may discover missing includes on systems where
socklen_t is available but the right headers are not included yet.
For example on win32 we need
  #ifdef HAVE_WS2TCPIP_H
  # include <ws2tcpip.h>
  #endif

Further patches should fix this. Here we do that for sanei/sanei_udp.c only
because this seems to be the only affected unit which would compile on win32
anyway.

Signed-off-by: Ruediger Meier <ruediger.meier@ga-group.nl>